+++ tests/data/cases/yield_singleton_tuple.py
@@ -1,0 +1,11 @@
+# flags: --preview
+
+def f():
+    yield x,
+
+
+# output
+
+
+def f():
+    yield (x,)